FinalProjectWIP_Build2
Michael Steele
4/20/17 V2:

- Started foundation work on the GUI. 
- Fixed a bug where the input of the seat letter would not be accepted if it was lower-case. 
(Wow I feel like a real developer typing in this format!)

The GUI appears close to Daniel's design but might need tweaking. I have initialized the JTextPane's value to display the airline seats immediately upon starting the program. I had to use JTextPane instead of JTextArea because the latter did not have a center-alignment function. I will double-check with the instructor next Tuesday to see if this is allowed.

If anyone would like to improve upon the appearance of the GUI I would be very grateful as I am still learning the ropes of Swing.