Build 3 - Final build MS
Everything works!
The GUI has been given full function and it will now act exactly how it should. The program is basically complete and meets all requirements for the final project sans the commenting. I've taken every single possible input into account when coding the input validation so the program's foolproof.
I just need to comment the code and it will be submit-ready. I know we wanted the Xs to be colored red for the pane but I haven't figured out how to pull that off. If you know you can do it then give it a go, it will only help us.

Thanks for all the help guys! It was a pretty fun project to work through. If you see any bugs or anything needing improvement make those changes and upload to this repository.